requirements.txt
Django>=4.2
psycopg2-binary
Pillow>=9.0
djangorestframework
django-cors-headers
python-dotenv
boto3>=1.34.0
python-multipart
requests

# Auth & Social login
djangorestframework-simplejwt
dj-rest-auth
django-allauth
python-social-auth
social-auth-app-django
google-auth
requests-oauthlib
